requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow>=2.15.0
pillow
numpy
python-dotenv